refactor(auth): add scope to token exchange and extract param interfaces
